Day2, Part 2: Pseudo implementation of recursive solution
This avoids having to re-evaluate the entire report multiple times, dropping a level each time. Instead, if we fail for the current level, we continue by ignoring it. We also remember that we skipped once, to not allow skipping multiple times.

This has better runtime complexity but for the input amounts given, it's not really relevant.