Day 12 part 2. Also added diagrams to explanation.
The second part was really hard for me. That might also be because I'm not at 100% brain capacity at the moment. It took me a while of experimenting to get the ordering of fences correct. I toyed with some other ideas that I have seen in other peoples solutions but couldn't make them work for me. Perhaps I'll revisit this when I'm fully healthy again....